updates per #37
From mubaris, points to be noted.

This is a web app. This is not a chrome extension. Chrome extension is based on logic of this app. That's completely different thing. You don't need to edit the Get Chrome Extension part.
Do not move badges to another section.
Rename Tools I used to Dependencies. This section is the list of libraries/frameworks/tools which are dependencies of Curiosity app.
We have 3 branches. gh-pages contains only front-end and it's hosted on Github Pages. v2 is the future of Curiosity. We are building a backend for this. The code for that is contained in v2. master is main branch. v2 will be merged to master once the process of building backend is done.
You need to change the screenshot of Curiosity in the Peek at Curiosity section. Upload the image to imgsafe then set the link.
We use Github labels and Github projects to manage tasks. up-for-grabs labels are the issues need to solved. discussion is for discussions. Check all labels and write about it. It's pretty straight forward.